Improve crash handling (#8)
* Set AFL_CMIN_ALLOW_ANY only for the first run, like afl-cmin.

Always setting AFL_CMIN_ALLOW_ANY includes crashes in the final output
of the cmin operation. This is unlike the original afl-cmin tool, which
only uses it for the first test run. It also prevents the -C flag from
functioning properly.

* Implement ability to filter out crashes into a separate directory.

Crashes are always deduplicated by hash (SHA1), even if --no-dedup is
passed.

* Mention -c in README.

* Don't dedup crash files if we previously did a global dedup.

* Use --crash-dir parameter name instead of -c.

* Ensure that we only modify copies of the env.

* Don't calculate crashed status inline.

* Clarify shutil.Error comment.

* If --crash-dir is not specified, treat crashes as ordinary inputs.

In other words, count them towards total coverage and preserve them as
part of the minimized corpus.

* Update README: `-c` -> `--crash-dir`